+++ next.config.ts
@@ -3,22 +3,24 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   reactStrictMode: true,
   experimental: {
-    serverActions: true,
+    serverActions: {
+      allowedOrigins: ["localhost:3000"],
+    },
+  },
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "i.ytimg.com",
+      },
+    ],
   },
   headers: async () => [
     {
       source: "/:path*",
-      headers: [
-        {
-          key: "X-Robots-Tag",
-          value: "noindex, nofollow",
-        },
-      ],
+      headers: [{ key: "X-Robots-Tag", value: "noindex, nofollow" }],
     },
   ],
-  images: {
-    domains: ["i.ytimg.com"],
-  },
 };
 
 export default nextConfig;